+++ src/BoardM5stackCore2.cpp
@@ -23,7 +23,6 @@ Overall, this initializes the specific hardware on the M5Stack Core2, configures
 #include "BoardInterface.h"
 #include "Board320_240.h"
 #include "BoardM5stackCore2.h"
-// #include "I2C_MPU6886.h"
 
 // Touch screen
 int16_t lastTouchX, lastTouchY;
@@ -34,19 +33,16 @@ bool btnAPressed = false;
 bool btnBPressed = false;
 bool btnCPressed = false;
 
-// I2C_MPU6886 imu(I2C_MPU6886_DEFAULT_ADDRESS, Wire1);
-
-/*float accX = 0.0F;  // Define variables for storing inertial sensor data
+float accX = 0.0F; // Define variables for storing inertial sensor data
 float accY = 0.0F;
 float accZ = 0.0F;
 float gyroX = 0.0F;
 float gyroY = 0.0F;
 float gyroZ = 0.0F;
 float pitch = 0.0F;
-float roll  = 0.0F;
-float yaw   = 0.0F;
+float roll = 0.0F;
+float yaw = 0.0F;
 float temp = 0.0F;
-*/
 
 /**
  * Init board
@@ -87,7 +83,7 @@ void BoardM5stackCore2::initBoard()
 
   M5.Touch.begin();
   M5.Rtc.begin();
-  // M5.IMU.Init(); // Gyro
+  M5.IMU.Init(); // Gyro
   // delay(100);
 
   Board320_240::initBoard();
@@ -348,28 +344,31 @@ void BoardM5stackCore2::boardLoop()
   M5.update();
   Board320_240::boardLoop();
 
-  /*    M5.IMU.getGyroData(&gyroX, &gyroY, &gyroZ);
-      M5.IMU.getAccelData(
-          &accX, &accY,
-          &accZ);
-      M5.IMU.getAhrsData(
-          &pitch, &roll,
-          &yaw);
-      M5.IMU.getTempData(&temp);  // Stores the inertial sensor temperature to
-      if (gyroX != 0.0 || gyroY != 0.0 || gyroZ != 0.0) {
-      syslog->printf("gyroX,  gyroY, gyroZ\n");
-      syslog->printf("%6.2f %6.2f%6.2f o/s\n", gyroX, gyroY, gyroZ);
-      }
-      if (accX != 0.0 || accY != 0.0 || accZ != 0.0) {
-      syslog->printf("accX,   accY,  accZ\n");
-      syslog->printf("%5.2f  %5.2f  %5.2f G\n", accX, accY, accZ);
-      }
-      if (pitch != 0.0 || roll != 0.0 || yaw != -8.5) {
-      syslog->printf("pitch,  roll,  yaw\n");
-      syslog->printf("%5.2f  %5.2f  %5.2f deg\n", pitch, roll, yaw);
-      }
-      syslog->flush();
-      */
+  M5.IMU.getGyroData(&gyroX, &gyroY, &gyroZ);
+  M5.IMU.getAccelData(
+      &accX, &accY,
+      &accZ);
+  M5.IMU.getAhrsData(
+      &pitch, &roll,
+      &yaw);
+  M5.IMU.getTempData(&temp);
+
+  if (gyroX != 0.0 || gyroY != 0.0 || gyroZ != 0.0)
+  {
+    syslog->printf("gyroX,  gyroY, gyroZ\n");
+    syslog->printf("%6.2f %6.2f%6.2f o/s\n", gyroX, gyroY, gyroZ);
+  }
+  if (accX != 0.0 || accY != 0.0 || accZ != 0.0)
+  {
+    syslog->printf("accX,   accY,  accZ\n");
+    syslog->printf("%5.2f  %5.2f  %5.2f G\n", accX, accY, accZ);
+  }
+  if (pitch != 0.0 || roll != 0.0 || yaw != -8.5)
+  {
+    syslog->printf("pitch,  roll,  yaw\n");
+    syslog->printf("%5.2f  %5.2f  %5.2f deg\n", pitch, roll, yaw);
+  }
+  syslog->flush();
 }
 
 /**