+++ src/BoardM5stackCore2.h
@@ -9,6 +9,7 @@
 
 #include "BoardInterface.h"
 #include "Board320_240.h"
+#include <M5Core2.h>
 
 class BoardM5stackCore2 : public Board320_240
 {