+++ src/LiveData.h
@@ -376,6 +376,8 @@ typedef struct
   uint8_t gpsModuleType;      // 0 - none, Default: 
   // 1 - u-blox NEO-M8N https://shop.m5stack.com/products/gps-module, 
   // 2 - GNSS Module https://shop.m5stack.com/products/gnss-module-with-barometric-pressure-imu-magnetometer-sensors
+  // == settings version 21
+  uint8_t carSpeedType;      // 0 - automatic (car/gps), 1 - only from car, 2 - only from gps
   //
 } SETTINGS_STRUC;